分別更改三大程序的程序名和图标
