--- app/rtkconv/rtkconv.cbproj.local
@@ -1,16 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<BorlandProject>
-	<Transactions>
-    <Transaction>2012/10/18 07:18:17.000.621,D:\proj\rtklib\app\rtkconv\Unit1.dfm=D:\proj\rtklib\app\rtkconv\codeopt.dfm</Transaction>
-    <Transaction>2012/10/18 07:18:17.000.621,D:\proj\rtklib\app\rtkconv\Unit1.h=D:\proj\rtklib\app\rtkconv\codeopt.h</Transaction>
-    <Transaction>2013/11/02 15:50:12.000.955,=D:\proj\RTKLIB\src\rcv\septentrio.c</Transaction>
-    <Transaction>2014/07/26 19:58:36.000.097,=D:\proj\RTKLIB\src\rcv\rcvlex.c</Transaction>
-    <Transaction>2014/08/26 08:52:10.000.736,=D:\proj\RTKLIB\src\rcv\rt17.c</Transaction>
-    <Transaction>2016/07/05 08:04:19.000.089,=D:\proj\RTKLIB\src\rcv\cmr.c</Transaction>
-    <Transaction>2017/05/26 18:40:37.000.629,=D:\proj\RTKLIB\src\rcv\tersus.c</Transaction>
-    <Transaction>2017/07/17 10:06:32.000.595,=C:\dev\RTKLIB\src\stream.c</Transaction>
-    <Transaction>2017/07/17 10:08:30.000.561,=C:\dev\RTKLIB\src\solution.c</Transaction>
-    <Transaction>2017/07/17 10:15:40.000.994,=C:\dev\RTKLIB\src\geoid.c</Transaction>
-    <Transaction>2017/07/17 10:15:47.000.075,=C:\dev\RTKLIB\src\erb.c</Transaction>
-  </Transactions>
-</BorlandProject>
